pyproject.toml
[project]
name = "ui-maker-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"google-genai>=0.7.0",
	"Pillow>=10.0.0",
	"pydantic>=2.0.0",
	"python-dotenv>=1.0.0"
]
